Start with google oauth2

+++ src/main/java/job_portal/feature/seeker/education/EducationServiceImpl.java
@@ -1,18 +1,17 @@
 package job_portal.feature.seeker.education;
-import java.time.LocalDate;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import job_portal.domain.backend.seeker.Degree;
 import job_portal.domain.backend.seeker.Education;
 import job_portal.domain.backend.seeker.Seeker;
-import job_portal.domain.backend.seeker.SeekerEducation;
 import job_portal.feature.seeker.auth.SeekerRepository;
 import job_portal.feature.seeker.degree.DegreeRepository;
 import job_portal.feature.seeker.education.dto.request.CreateEductionRequest;
 import job_portal.feature.seeker.education.dto.request.UpdateEducationRequest;
 import job_portal.feature.seeker.education.dto.respone.EducationRespone;
 import job_portal.mapper.seeker.EducationMapper;
+import job_portal.security.JwtService;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -23,21 +22,32 @@ public class EducationServiceImpl implements EducationService{
     private final DegreeRepository degreeRepository;
     private final EducationMapper educationMapper;
     private final EducationRepository educationRepository;
-    private final SeekerEducationRepository seekerEducationRepository;
+    private final JwtService jwtService;
+    // private final SeekerEducationRepository seekerEducationRepository;
 
     @Override
     public void deleteById(Integer id) {
+        Long seekerId = jwtService.extractUserId();
         Education education = educationRepository.findById(id).orElseThrow(
             ()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Education not found!")
         );
+        if (!education.getSeeker().getId().equals(seekerId.intValue())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                "You do not update this education");
+        }
         educationRepository.delete(education);
     }   
 
     @Override
     public EducationRespone updateById(Integer id, UpdateEducationRequest updateEducationRequest) {
+        Long seekerId = jwtService.extractUserId();
         Education education = educationRepository.findById(id).orElseThrow(
             ()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Education not found!")
         );
+        if (!education.getSeeker().getId().equals(seekerId.intValue())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                "You do not update this education");
+        }
         educationMapper.fromEducationUpdateRequest(updateEducationRequest, education);
         education = educationRepository.save(education);
         return educationMapper.toEducationRespone(education);
@@ -54,15 +64,8 @@ public class EducationServiceImpl implements EducationService{
         
         Education education = educationMapper.fromCreateNewEducation(createEductionRequest);
         education.setDegree(degree);
-        // Save to database Education
+        education.setSeeker(seeker);
+
         educationRepository.save(education);
-
-        SeekerEducation seekerEducation = new SeekerEducation();
-        seekerEducation.setSeeker(seeker);
-        seekerEducation.setEducation(education);
-        seekerEducation.setCreatedAt(LocalDate.now());
-
-        // Save to database SeekerEducation
-        seekerEducationRepository.save(seekerEducation);
-    } 
+    }
 }
